Add files via upload
The current antivirus system is designed to perform fundamental security functions such as scanning for malware, monitoring system performance, and optimizing RAM usage. It operates primarily based on predefined malware signatures, which are stored in a local database. The scanning process is initiated manually by the user, and the system provides basic UI elements for interaction. While it serves as a functional security tool, it lacks the advanced automation and integration features found in modern antivirus solutions. Additionally, it does not offer real-time protection, making it less effective against emerging threats


•	Python Official Documentation: Served as a primary resource for implementing core functionalities, including file scanning, hashing, and process management.
•	CustomTkinter Library: Assisted in designing a modern and responsive user interface, improving the overall user experience.
•	Pygame & Psutil: Pygame was used for UI responsiveness, while Psutil enabled real-time system monitoring, CPU usage tracking, and performance optimization.
•	VirusTotal API: Provided cloud-based malware scanning capabilities, allowing the system to cross-check files with a vast online database of known threats.
•	Scikit-learn: Supported AI-driven threat analysis, enabling the integration of machine learning techniques for improved malware detection.